refactor: `rawdb` tests
